refactor(api): move obfuscation-service from shared to iam context
